added more cli actions
