+++ src-tauri/src/main.rs
@@ -24,6 +24,7 @@ use commands::{
 };
 use logging::logger;
 use migrations::{Migrator, MigratorTrait};
+use crate::cli::cli;
 
 mod db;
 mod commands;
@@ -32,6 +33,7 @@ mod cache;
 mod config;
 mod utils;
 mod logging;
+mod cli;
 
 #[tokio::main]
 async fn main() {
@@ -52,15 +54,7 @@ async fn main() {
 			.map_err(|e| {
 				log::error!("{:?}", e);
 			}).expect("Error connecting to DB");
-	
-	// TODO move into setup
-	Migrator::up(&connection, None)
-			.await
-			.into_report()
-			.attach_printable("Error running migrations")
-			.map_err(|e| {
-				log::error!("{:?}", e);
-			}).expect("Error running migrations");
+	let conn = connection.clone();
 	
 	let cache = Mutex::new(cache::create()
 			.map_err(|e| {
@@ -82,86 +76,24 @@ async fn main() {
 				}
 				Ok(())
 			})
+			.setup(|_| {
+				let (tx, rx) = channel();
+				tauri::async_runtime::spawn(async move {
+					Migrator::up(&conn, None)
+							.await
+							.into_report()
+							.attach_printable("Error running migrations")
+							.map_err(|e| {
+								log::error!("{:?}", e);
+							}).expect("Error running migrations");
+					tx.send(()).expect("Error sending message (update)");
+				});
+				
+				rx.recv().expect("Error receiving message (update)");
+				Ok(())
+			})
 			.setup(|app| {
-				if let Ok(matches) = app.get_cli_matches() {
-					println!();
-					let mut terminate_after_cli = true;
-					
-					println!("CLI args: {:?}", matches.args);
-					
-					for m in matches.args {
-						match (m.0.as_ref(), m.1) {
-							("update", d) => {
-								if let Value::Bool(b) = d.value {
-									if b {
-										let (tx, rx) = channel();
-										let handle = app.handle();
-										tauri::async_runtime::spawn(async move {
-											match handle.updater().check().await {
-												Ok(update) => {
-													if update.is_update_available() {
-														log::info!("Update available: {}", update.latest_version());
-														if let Err(e) = update.download_and_install().await {
-															log::error!("Error downloading update: {}", e);
-														} else {
-															log::info!("Update successfully");
-														}
-													} else {
-														log::error!("No update available");
-													}
-												}
-												Err(e) => {
-													log::error!("Error checking for update: {}", e);
-												}
-											}
-											tx.send(()).expect("Error sending message (update)");
-										});
-										
-										rx.recv().expect("Error receiving message (update)");
-									}
-								}
-							}
-							("help", d) => {
-								println!("{}", d.value.as_str().expect("help value is not a string"));
-							}
-							("version", d) => {
-//								if let Value::Bool(b) = d.value {
-								// tauri filters all other args, and sets this to Null if -v is passed
-								if let Value::Null = d.value {
-//									if b {
-										println!("{}: {}", built_info::PKG_NAME, built_info::PKG_VERSION);
-//									}
-								}
-							}
-							("config", d) => {
-								if let Value::Bool(b) = d.value {
-									if b {
-										println!("{}", dirs::create_conf_folder().expect("Error creating config folder").display());
-									}
-								}
-							}
-							("data", d) => {
-								if let Value::Bool(b) = d.value {
-									if b {
-										println!("{}", dirs::create_data_folder().expect("Error creating data folder").display());
-									}
-								}
-							}
-							("run", d) => {
-								if let Value::Bool(b) = d.value {
-									if b {
-										terminate_after_cli = false;
-									}
-								}
-							}
-							_ => {}
-						};
-					}
-					if terminate_after_cli {
-						let handle = app.handle();
-						handle.exit(0)
-					}
-				}
+				cli(app);
 				Ok(())
 			})
 			.manage(connection)
